Не сделал только ообработку сценария с отсутствием описания, help
